rewrite the C version, the previous version is too keen on the code safety.
I made an attempt to the previous version, so it keen on the code safety too much, that's not a good idea. I modify it, add my own assert, my own malloc, and just make the whole program much easier to understand. Now, the file to record the file attributes list will record the file last write time too.